@@ -5,7 +5,7 @@ from pydantic import BaseModel
 
 class HashableModel(BaseModel):
     def __hash__(self) -> int:
-        return hash(self.json())
+        return hash(self.model_dump_json())
 
 
 class Sort(str, Enum):
